Unified check-rule-3.x.py scripts with a single checklib.py executable
